requirements.txt
pandas
streamlit
plotly
python-dotenv
altair
bokeh==2.4.1
requests
seaborn
